pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sqloptimenv"
version = "1.0.0"
description = "OpenEnv environment for SQL query optimization"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "pydantic>=2.7.4",
    "openai>=1.35.13",
    "requests>=2.32.3",
    "pyyaml>=6.0.2",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["sqloptimenv*", "server*"]
